+++ src/wrappers.rs
@@ -1,4 +1,4 @@
-//! A set of wrappers around the MuJoCo types.
+//! A set of wrappers around the MuJoCo types and functions.
 pub mod mj_visualization;
 pub mod mj_interface;
 pub mod mj_statistic;